+++ migrations/20260119030543_create_users.sql
@@ -1,9 +1,9 @@
 -- up
-CREATE TABLE users (
+CREATE TABLE premix_migration_test_users (
     id INTEGER PRIMARY KEY,
     username TEXT NOT NULL,
     email TEXT NOT NULL
 );
 
 -- down
-DROP TABLE users;
+DROP TABLE premix_migration_test_users;