+++ orm-book/book-examples/tests/multi_db.rs
@@ -1,0 +1,4 @@
+#[test]
+fn sqlite_pool_type_compiles() {
+    let _ = std::any::TypeId::of::<premix_orm::sqlx::SqlitePool>();
+}